Refector : 공고 신청, 개인정보 수정, 개인정보 조회 및 마이페이지 공고 관련 API Refectoring

+++ src/main/java/com/archisemtle/semtlewebserverspring/application/ApplyProjectService.java
@@ -4,5 +4,5 @@ import com.archisemtle.semtlewebserverspring.dto.ApplyProjectRequestDto;
 import com.archisemtle.semtlewebserverspring.dto.ApplyProjectResponseDto;
 
 public interface ApplyProjectService {
-    ApplyProjectResponseDto applyProject(Integer boardId, Integer applicantId, ApplyProjectRequestDto applyProjectRequestDto) throws Exception;
+    ApplyProjectResponseDto applyProject(Long boardId, Long applicantId, ApplyProjectRequestDto applyProjectRequestDto) throws Exception;
 }
